requirements.txt
requests
python-dotenv
streamlit
pymongo
